Add importing the current existing list of Pokemon (which will be updated as time goes on) and added a function which allows checking type coverage for a given move pool.
There is an unused flag (that will be coded in soon as I figure out how I
want to structure abilities for easy, non-technical user modifications)
that allows considering abilities on Pokemon when doing the type effectiveness
grouping. See http://pokemondb.net/tools/type-coverage for the original implementation.
This would be useful for people designing fakemon to make sure their movepool possibilities
are as (im)balanced as intended.